added ints instead of longs

+++ src/main/java/org/piki/App.java
@@ -11,9 +11,9 @@ public class App {
     public static long fibonacci(int n) {
         if (n <= 2) return 1;
 
-        long a = 1, b = 1;
+        int a = 1, b = 1;
         for (int i=2; i<n; i++) {
-            long t = a;
+            int t = a;
             a = b;
             b = b + t;
         }
